+++ de.dlr.sc.virsat.model.extension.mechanical.ui/src/de/dlr/sc/virsat/model/extension/mechanical/ui/wizards/CatiaExportPage.java
@@ -10,18 +10,9 @@
 package de.dlr.sc.virsat.model.extension.mechanical.ui.wizards;
 
 import org.eclipse.core.resources.IContainer;
-import org.eclipse.jface.viewers.TreeViewer;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.widgets.Composite;
-import org.eclipse.swt.widgets.FileDialog;
 import org.eclipse.swt.widgets.Label;
-
-import de.dlr.sc.virsat.model.extension.ps.model.AssemblyTree;
-import de.dlr.sc.virsat.model.extension.ps.model.ConfigurationTree;
-import de.dlr.sc.virsat.model.extension.ps.model.ElementConfiguration;
-import de.dlr.sc.virsat.model.extension.ps.model.ElementOccurence;
-import de.dlr.sc.virsat.project.ui.contentProvider.VirSatFilteredWrappedTreeContentProvider;
-import de.dlr.sc.virsat.uiengine.ui.wizard.AImportExportPage;
 
 /**
  * This class configures the export page by providing the selection
@@ -30,10 +21,7 @@ import de.dlr.sc.virsat.uiengine.ui.wizard.AImportExportPage;
  *
  */
 
-public class CatiaExportPage extends AImportExportPage {
-
-	private static final String DIALOG_TEXT = "File name";
-	private static final String[] DIALOG_EXTENSIONS = { "*.json" };
+public class CatiaExportPage extends ACatiaImportExportPage {
 
 	/**
 	 * Standard constructor
@@ -50,24 +38,13 @@ public class CatiaExportPage extends AImportExportPage {
 	@Override
 	public void createControl(Composite parent) {
 		super.createControl(parent);
-		TreeViewer treeViewer = createTreeUI();
-		VirSatFilteredWrappedTreeContentProvider filteredCp = (VirSatFilteredWrappedTreeContentProvider) treeViewer.getContentProvider();
-		filteredCp.addStructuralElementIdFilter(ConfigurationTree.FULL_QUALIFIED_STRUCTURAL_ELEMENT_NAME);
-		filteredCp.addStructuralElementIdFilter(ElementConfiguration.FULL_QUALIFIED_STRUCTURAL_ELEMENT_NAME);
-		filteredCp.addStructuralElementIdFilter(AssemblyTree.FULL_QUALIFIED_STRUCTURAL_ELEMENT_NAME);
-		filteredCp.addStructuralElementIdFilter(ElementOccurence.FULL_QUALIFIED_STRUCTURAL_ELEMENT_NAME);
 		
+		createTreeViewer();
 		createFileDestinationUI();
 		
 		Label label = new Label((Composite) getControl(), SWT.NONE);
 		label.setText("All attached Geometry files will be copied into the same directory as the JSON file.");
 	}
 	
-	@Override
-	protected String openDialog() {
-		FileDialog dialog = new FileDialog(getContainer().getShell(), SWT.SAVE | SWT.SHEET);
-		dialog.setText(DIALOG_TEXT);
-		dialog.setFilterExtensions(DIALOG_EXTENSIONS);
-		return dialog.open();
-	}
+
 }
